fix(engine): int32_t/short for tempsectorz/tempsectorpicnum arrays
source/GAME.C:2666-2667 declared these as 'long' but they shadow
sector[].floorz/ceilingz (int32_t per BUILD.H:56) and
sector[].floorpicnum/ceilingpicnum (short per BUILD.H:60). Match the
field types directly to eliminate the 64-bit type mismatch in
SE40_Draw floor-over-floor effects.

Only caller is SE40_Draw in this same translation unit; no callers
outside source/GAME.C (grep -rn confirmed).

Closes: fix-engine-tempsectorz-type-mismatch

+++ source/GAME.C
@@ -2663,8 +2663,8 @@ void drawbackground(void)
 #define FOFTILE 13
 #define FOFTILEX 32
 #define FOFTILEY 32
-long tempsectorz[MAXSECTORS];
-long tempsectorpicnum[MAXSECTORS];
+int32_t tempsectorz[MAXSECTORS];
+short tempsectorpicnum[MAXSECTORS];
 //short tempcursectnum;
 
 SE40_Draw(int spnum,long x,long y,long z,short a,short h,long smoothratio)
